Introduce `ChannelExclusionFormBuilder` for managing excluded channels in promotion forms and refactor related extensions and tests.
